Renamed alkymi.py to core.py to give it a more meaningful name
