+++ tests/test_utils.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 import io
-import os
 import subprocess
 import threading
 import time
